Quick fix to invite not deleting bug
NOTE: noticed a memory leak in firebase, with data constantly being updated/read for no reason. After refreshing the page and trying to recreate the issue 5 times, I could not. I think this should be a safe commit now but I'm leaving this as a warning, because the memory leak only stopped upon closing the app, which could lead to lots of charges to our account.